+++ wsgidav/dc/simple_dc.py
@@ -7,20 +7,50 @@
 Implementation of a domain controller that uses realm/user_name/password mappings
 from the configuration file and uses the share path as realm name.
 
-user_map is defined a follows::
+user_mapping is defined a follows::
 
-    user_map = {'realm1': {
-                 'John Smith': {'description': '',
-                                'password': 'YouNeverGuessMe',
-                                },
-                 'Dan Brown': {'description': '',
-                               'password': 'DontGuessMeEither',
-                               },
-                 }
-               'realm2': {
-                 ...
-                 }
-               }
+    simple_dc: {
+        user_mapping = {
+            "realm1": {
+                "John Smith": {
+                    "password": "YouNeverGuessMe",
+                },
+                "Dan Brown": {
+                    "password": "DontGuessMeEither",
+                    "roles": ["editor"]
+                }
+            },
+            "realm2": {
+                ...
+            }
+        },
+    }
+
+The "*" pseudo-share is used to pass a default definition::
+
+    user_mapping = {
+        "*": {  // every share except for 'realm2'
+            "Dan Brown": {
+                "password": "DontGuessMeEither",
+                "roles": ["editor"]
+            }
+        },
+        "realm2": {
+            ...
+        }
+    },
+
+A share (even the "*" pseudo-share) can be set to True to allow anonymous access::
+
+    user_mapping = {
+        "*": {
+            "Dan Brown": {
+                "password": "DontGuessMeEither",
+                "roles": ["editor"]
+            },
+        },
+        "realm2": True
+    },
 
 The SimpleDomainController fulfills the requirements of a DomainController as
 used for authentication with http_authenticator.HTTPAuthenticator for the
@@ -49,25 +79,45 @@ class SimpleDomainController(BaseDomainController):
         self.user_map = dc_conf.get("user_mapping")
         if self.user_map is None:
             raise RuntimeError("Missing option: simple_dc.user_mapping")
+        for share, data in self.user_map.items():
+            if type(data) not in (bool, dict) or not data:
+                raise RuntimeError(
+                    "Invalid option: simple_dc.user_mapping['{}']: must be True or non-empty dict.".format(
+                        share
+                    )
+                )
+        return
 
     def __str__(self):
         return "{}()".format(self.__class__.__name__)
 
+    def _get_realm_entry(self, realm, user_name=None):
+        """Return the matching user_map entry (falling back to default '*' if any)."""
+        realm_entry = self.user_map.get(realm)
+        if realm_entry is None:
+            realm_entry = self.user_map.get("*")
+        if user_name is None:
+            return realm_entry
+        return realm_entry.get(user_name)
+
     def get_domain_realm(self, path_info, environ):
-        """Resolve a relative url to the  appropriate realm name."""
-        realm = self._calc_realm_from_path(path_info, environ)
+        """Resolve a relative url to the appropriate realm name."""
+        realm = self._calc_realm_from_path_provider(path_info, environ)
         return realm
 
     def require_authentication(self, realm, environ):
-        """Return True if this realm requires authentication or False if it is
-        available for general access."""
-        # TODO: The user_mapping syntax should make this more explicit!
-        return realm in self.user_map
+        """Return True if this realm requires authentication (grant anonymous access otherwise)."""
+        realm_entry = self._get_realm_entry(realm)
+        if realm_entry is None:
+            _logger.error("No user_mapping or '*' defined for realm '{}'".format(realm))
+        return realm_entry is not True
+        # return realm not in self.user_map
 
     def basic_auth_user(self, realm, user_name, password, environ):
         """Returns True if this user_name/password pair is valid for the realm,
         False otherwise. Used for basic authentication."""
-        user = self.user_map.get(realm, {}).get(user_name)
+        # user = self.user_map.get(realm, {}).get(user_name)
+        user = self._get_realm_entry(realm, user_name)
 
         if user is not None and password == user.get("password"):
             environ["wsgidav.auth.roles"] = user.get("roles", [])
@@ -81,5 +131,8 @@ class SimpleDomainController(BaseDomainController):
 
     def digest_auth_user(self, realm, user_name, environ):
         """Computes digest hash A1 part."""
-        password = self.user_map.get(realm, {}).get(user_name, {}).get("password")
+        user = self._get_realm_entry(realm, user_name)
+        password = user.get("password")
+        # password = self.user_map.get(realm, {}).get(user_name, {}).get("password")
+        environ["wsgidav.auth.roles"] = user.get("roles", [])
         return self._compute_http_digest_a1(realm, user_name, password)